Minor restructure: moved Commands enum and match- expression into commands/mod.rs, in line with the structure of the other subcommands.
